+++ R/getContrXlev.R
@@ -2,7 +2,7 @@
 ### R routines for the R package mixmeta (c)
 #
 getContrXlev <-
-function(formula, list)  {
+function(terms, list)  {
 #
 ################################################################################
 # FUNCTION TO EXTRACT THE CONTRASTS/LEVELS RELATED TO A GIVEN FORMULA
@@ -11,7 +11,7 @@ function(formula, list)  {
   if(is.null(list)) return(NULL)
 #
   # RETURN CONSTRASTS RELATED TO TERMS IN THE FORMULA
-  vars <- vapply(attr(terms(formula), "variables"), deparse, "")[-1L]
+  vars <- vapply(attr(terms, "variables"), deparse, "")[-1L]
   ind <- names(list) %in% vars
   if(any(ind)) list[ind] else NULL
 }