+++ R/getFullformula.R
@@ -2,7 +2,7 @@
 ### R routines for the R package mixmeta (c)
 #
 getFullFormula <-
-function(formula, random)  {
+function(formula, random, data=NULL)  {
 #
 ################################################################################
 # FUNCTION TO ADD RANDOM TERMS (PREDICTORS AND GROUPING VARS) IN FORMULA
@@ -12,10 +12,11 @@ function(formula, random)  {
 #
   # EXTRACT THE TERMS IN FORMULAE FOR FIXED AND RANDOM TERMS
   random <- getList(random)
-  fixterms <- attr(terms(formula),"term.labels")
+  fixterms <- attr(terms(formula, data=data), "term.labels")
   modrandom <- lapply(random, function(x)
     formula(paste("~",gsub("|","+",deparse(x[[2]]),fixed=TRUE))))
-  ranterms <- unlist(lapply(modrandom, function(x) attr(terms(x),"term.labels")))
+  ranterms <- unlist(lapply(modrandom, function(x) 
+    attr(terms(x, data=data),"term.labels")))
 #
   # ADD (MISSING) RANDOM TERMS TO FORMULA
   add <- unique(ranterms[!ranterms%in%fixterms])